apps/api/requirements.txt
fastapi==0.100.0
langchain==0.0.264
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.18
psycopg2-binary==2.9.6
pinecone-client==2.2.2
pdfminer.six==20221105
tiktoken==0.4.0
openai==0.27.8
lark==1.1.7
PyPDF2==3.0.1
unstructured==0.9.0
pdf2image==1.16.3
jinja2==3.1.2